disp: msm: sde: avoid encoder power-collapse with pending frames
The encoder idle work is scheduled during the frame-done event
to be executed after a timeout. During the execution, the check
for any on-going/pending frames is invalid as it checks for > 1.
Fix it to check for any non-zero frame-pending and avoid
power-collapse.

Change-Id: If7715ee56cc9bfa63787811458f4fc91de540013
Signed-off-by: Veera Sundaram Sankaran <veeras@codeaurora.org>
Signed-off-by: Ray Zhang <rayz@codeaurora.org>
Signed-off-by: Bruce Hoo <bingchua@codeaurora.org>
(cherry picked from commit 6c6d649c3ea0d59250ab22b3dec383f7f57ad976)